Move GSModel JUnit tests to new sub-package
